backend/requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# HTTP Client
httpx==0.26.0

# Crypto
pycryptodome==3.19.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Utils
python-dotenv==1.0.0
